fix(stock_notification_service) Fixing the Kafka handle_sneaker_sizes event handler
- Added an explicit event type check for calling the delete_sneaker_association coroutine

+++ backend/stock_notification_service/stock_notification_service/stock_notification/kafka/kafka_handlers/sneaker_sizes_handler.py
@@ -26,7 +26,7 @@ async def handle_sneaker_sizes(key: str | None, value: dict):
                 data = value.get("data")
                 sneaker_sizes_update = SneakerSizeUpdate(**data)
                 await update_sneaker_sizes(session, int(key), sneaker_sizes_update)
-            else:
+            elif event_type == "sneaker_sizes_deleted":
                 data = value.get("data")
                 sneaker_assoc_delete = SneakerAssocsDelete(**data)
                 await delete_sneaker_association(
